+++ CDF3/src/libf/get_var_realf.c
@@ -21,7 +21,7 @@
 
 /* Prototypes for the Fortran interfaces */
 #include "mpifnetcdf.h"
-FORTRAN_API int FORT_CALL nfmpi_get_var_real_ ( int *v1, int64_t *v2, float*v3 ){
+FORTRAN_API int FORT_CALL nfmpi_get_var_real_ ( int *v1, int *v2, float*v3 ){
     int ierr;
     int l2 = *v2 - 1;
     ierr = ncmpi_get_var_float( *v1, l2, v3 );
